+++ Emuera/GlobalStatic.cs
@@ -56,6 +56,9 @@ internal static class GlobalStatic
 	#endregion
 
 	public static CtrlZ ctrlZ = new();
+	#region EE_CALLSHARP注意
+	public static bool ExistPlugin;
+	#endregion
 
 #if DEBUG
 	public static List<FunctionLabelLine> StackList = [];