+++ Emuera/Runtime/Config/Config.cs
@@ -147,6 +147,9 @@ internal static class Config
 		#region EE_行連結の改行コード置換
 		ReplaceContinuationBR = instance.GetConfigValue<string>(ConfigCode.ReplaceContinuationBR);
 		#endregion
+		#region EE_CALLSHARP注意
+		PluginAvailableWarn = instance.GetConfigValue<bool>(ConfigCode.PluginAvailableWarn);
+		#endregion
 
 		#region EM_私家版_LoadText＆SaveText機能拡張
 		ValidExtension = instance.GetConfigValue<List<string>>(ConfigCode.ValidExtension);
@@ -700,6 +703,9 @@ internal static class Config
 	#endregion
 	#region EE_行連結の改行コード置換
 	public static string ReplaceContinuationBR { get; internal set; }
+	#endregion
+	#region EE_CALLSHARP注意
+	public static bool PluginAvailableWarn { get; private set; }
 	#endregion
 	#region EM_私家版_LoadText＆SaveText機能拡張
 	public static List<string> ValidExtension { get; private set; }