+++ Emuera/Runtime/Utils/EvilMask/Lang.cs
@@ -160,6 +160,7 @@ internal sealed partial class Lang
 				[Managed] public static TranslatableString UseSaveFolder { get; } = new TranslatableString("セーブデータをsavフォルダ内に作成する");
 				[Managed] public static TranslatableString EnglishConfigOutput { get; } = new TranslatableString("CONFIGファイルの内容を英語で保存する");
 				[Managed] public static TranslatableString UseCtrlZUndo { get; } = new TranslatableString("Zで元に戻す機能を有効にする");
+				[Managed] public static TranslatableString PluginAvailableWarn { get; } = new TranslatableString("外部プラグインが有効時に警告を表示する");
 				[Managed] public static TranslatableString MaxLog { get; } = new TranslatableString("履歴ログの行数");
 				[Managed] public static TranslatableString InfiniteLoopAlertTime { get; } = new TranslatableString("無限ループ警告までのミリ秒");
 				[Managed] public static TranslatableString SaveDataPerPage { get; } = new TranslatableString("使用するセーブデータ数");
@@ -1342,7 +1343,8 @@ internal sealed partial class Lang
 		[Managed] public static TranslatableString CreateFromCSV { get; } = new TranslatableString("ファイル「{0}」からスプライト「{1}」を生成・・・");
 		[Managed] public static TranslatableString Variant { get; } = new TranslatableString("# バリアント");
 		[Managed] public static TranslatableString NotDefinedGameBase { get; } = new TranslatableString("GameBase未定義");
-		
+		[Managed] public static TranslatableString PluginAvailable { get; } = new TranslatableString("注意：外部プラグイン機能が有効になっています。この機能で生じた不具合等はEmueraのサポート対象外となります");
+
 		#region LazyLoading
         [Managed] public static TranslatableString LazyLoadingNoConfigFile { get; } = new TranslatableString("遅延ローディング設定ファイルが見つからないため、テーブル構築をスキップします");
         [Managed] public static TranslatableString LazyLoadingConfigError { get; } = new TranslatableString("遅延ローディング設定ファイルの読み込みに失敗しました。メッセージ：{0}");